+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.8] - 2024-11-28
+
+### Changed
+- **CSIPVocabularyContentCategory** resource added, function for using dictionary and or term list added
+- **CSIPVocabularyContentInformationType** resource added, function for using dictionary and or term list added
+
 ## [0.2.7] - 2024-11-25
 ### Changed
 - **Pairtree** as directory-based storage with differential versioning.